requirement.txt
pycryptodome
opencv-python
mysql-connector-python
pytesseract
django